Add more tests for upgraded macro tag, fix usage with args
