+++ docs/core-tags.md
@@ -302,27 +302,40 @@ Special HTML characters will _not_ be escaped since the file is expected to be a
 ### `<macro>`
 
 Parameterized macros allow for reusable fragments within an HTML template.
-A macro can be defined using the `<macro>` directive.
+A macro can be defined using the `<macro>` tag.
 
 ```marko
-<macro greeting(name, count)>
+<macro({ name, count }) name="greeting">
     Hello ${name}! You have ${count} new messages.
 </macro>
 ```
 
-The above macro can then be invoked as part of any expression. The following
-sample template shows how to use macro functions inside expressions:
+The above macro can then be used as if it was a regular `greeting` tag. The following
+sample template shows how to use macro functions inside a template:
 
 ```marko
-<macro greeting(name, count)>
+<macro({ name, count }) name="greeting">
     Hello ${name}! You have ${count} new messages.
 </macro>
+
 <p>
-    <greeting("John", 10)/>
-</p>
-<p>
-    <!-- Or, using named attributes: -->
     <greeting name="Frank" count=20/>
+</p>
+```
+
+Macro's receive input similar to the root template, including a `renderBody` for displaying any provided body content like so:
+
+```marko
+<macro({ renderBody }) name="special-heading">
+    <h1>
+        <${renderBody}/>!
+    </h1>
+</macro>
+
+<p>
+    <special-heading>
+        Hello!
+    </special-heading>
 </p>
 ```
