+++ test/migrate/fixtures/legacy-macro-syntax/snapshot-expected.marko
@@ -7,3 +7,9 @@
 </macro>
 <test firstName="John" lastName="Smith"/>
 <test firstName="John" lastName="Smith"/>
+<macro(macroInput) name="test2">
+    <div>
+        <${macroInput}/>
+    </div>
+</macro>
+<test2>Hello</test2>
