+++ test/migrate/fixtures/legacy-macro-syntax/template.marko
@@ -8,3 +8,14 @@
 <test("John", "Smith")/>
 <test firstName="John" lastName="Smith"/>
 
+<macro test2()>
+  <div>
+    <macro-body/>
+  </div>
+</macro>
+
+
+<test2>
+  Hello
+</test2>
+
