+++ test/render/fixtures-deprecated/macro-body/expected.html
@@ -1,0 +1,1 @@
+<div><h1>My Alert</h1><p>Something went wrong!</p></div>
