+++ test/render/fixtures-deprecated/macro-body/test.js
@@ -1,0 +1,1 @@
+exports.templateData = {};
