+++ test/render/fixtures-deprecated/macro-boolean/expected.html
@@ -1,0 +1,1 @@
+<div>Hello Frank! You are 30 years old.</div><div class="hidden">Hello John! You are 10 years old.</div><div class="hidden">Hello John! You are 10 years old.</div><div>Hello John! You are 10 years old.</div>
