+++ test/render/fixtures-deprecated/macro-boolean/template.marko
@@ -1,4 +1,4 @@
-<macro({ name, age, hidden }) name="greeting">
+<macro greeting(name, age, hidden)>
     <div class=(hidden === true ? 'hidden' : null)>
         Hello ${name}! You are ${age} years old.
     </div>
