+++ test/render/fixtures-deprecated/macro-boolean/test.js
@@ -1,0 +1,1 @@
+exports.templateData = {};
