+++ test/render/fixtures-deprecated/macro-no-args/expected.html
@@ -1,0 +1,1 @@
+Marko is awesomeMarko is awesomeMarko is really awesomeMarko is really awesome
