+++ test/render/fixtures-deprecated/macro-no-args/template.marko
@@ -1,8 +1,8 @@
-<macro name="greeting">
+<macro greeting()>
     Marko is awesome
 </macro>
 
-<macro name="greeting2">
+<macro greeting2>
     Marko is really awesome
 </macro>
