+++ test/render/fixtures-deprecated/macro-positional-args/expected.html
@@ -1,0 +1,1 @@
+Hello Frank! You are 30 years old.Hello Doe, John! You are 10 years old.
