+++ test/render/fixtures-deprecated/macro-positional-args/template.marko
@@ -1,4 +1,4 @@
-<macro(name, age) name="greeting">
+<macro greeting(name, age)>
   Hello ${name}! You are ${age} years old.
 </macro>
