+++ test/render/fixtures-deprecated/macro-positional-args/test.js
@@ -1,0 +1,1 @@
+exports.templateData = {};
