+++ test/render/fixtures-deprecated/macro/expected.html
@@ -1,0 +1,1 @@
+Hello Frank! You are 30 years old.Hello John! You are 10 years old.
