+++ test/render/fixtures-deprecated/macro/test.js
@@ -1,0 +1,1 @@
+exports.templateData = {};
