+++ test/render/fixtures-deprecated/macros-ordering/expected.html
@@ -1,0 +1,1 @@
+[a: Hello foo] [b: Hello bar]
