+++ test/render/fixtures-deprecated/macros-ordering/template.marko
@@ -1,9 +1,9 @@
-<macro(name) name="a">
+<macro a(name)>
     [a: Hello ${name}]
     <b('bar')/>
 </macro>
 
-<macro(name) name="b">
+<macro b(name)>
     [b: Hello ${name}]
 </macro>
