+++ test/render/fixtures-deprecated/macros-ordering/test.js
@@ -1,0 +1,1 @@
+exports.templateData = {};
