+++ test/render/fixtures/text-replacement/expected.html
@@ -1,1 +1,1 @@
- Hello John. You are from San Jose, CA Zero: 0
+Hello John. You are from San Jose, CA Zero: 0
